+++ web/src/main/java/com/github/freeacs/web/App.java
@@ -3,18 +3,15 @@ package com.github.freeacs.web;
 import com.github.freeacs.web.app.Main;
 import com.github.freeacs.web.app.Monitor;
 import com.github.freeacs.web.app.menu.MenuServlet;
-import com.github.freeacs.web.app.security.LoginServlet;
 import com.github.freeacs.web.app.util.Freemarker;
 import com.github.freeacs.web.help.HelpServlet;
 import com.zaxxer.hikari.HikariDataSource;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.boot.autoconfigure.flyway.FlywayAutoConfiguration;
 import org.springframework.boot.context.properties.ConfigurationProperties;
 import org.springframework.boot.jdbc.DataSourceBuilder;
-import org.springframework.boot.web.servlet.FilterRegistrationBean;
 import org.springframework.boot.web.servlet.ServletRegistrationBean;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Primary;
@@ -60,7 +57,7 @@ public class App {
         ServletRegistrationBean<Main> srb = new ServletRegistrationBean<>();
         srb.setServlet(new Main(mainDataSource, syslogDataSource));
         srb.setName("main");
-        srb.setUrlMappings(Collections.singletonList("/web"));
+        srb.setUrlMappings(Collections.singletonList(Main.servletMapping));
         return srb;
     }
