+++ web/src/main/java/com/github/freeacs/web/app/Main.java
@@ -5,7 +5,6 @@ import com.github.freeacs.web.Page;
 import com.github.freeacs.web.app.input.ParameterParser;
 import com.github.freeacs.web.app.menu.MenuServlet;
 import com.github.freeacs.web.app.page.WebPage;
-import com.github.freeacs.web.app.security.WebUser;
 import com.github.freeacs.web.app.util.Freemarker;
 import com.github.freeacs.web.app.util.SessionCache;
 import com.github.freeacs.web.app.util.WebProperties;
@@ -13,7 +12,6 @@ import freemarker.template.Configuration;
 import freemarker.template.TemplateException;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import org.springframework.security.core.context.SecurityContextHolder;
 
 import javax.servlet.ServletException;
 import javax.servlet.http.HttpServlet;
@@ -39,6 +37,8 @@ public class Main extends HttpServlet {
 
 	/** The Constant version. */
 	public static final String version = "2.2.58";
+
+	public static final String servletMapping = "/web";
 
 	/** The config. */
 	private Configuration config;
