+++ web/src/main/java/com/github/freeacs/web/app/Output.java
@@ -8,10 +8,10 @@ import com.github.freeacs.web.app.input.InputSelectionFactory;
 import com.github.freeacs.web.app.input.ParameterParser;
 import com.github.freeacs.web.app.menu.MenuItem;
 import com.github.freeacs.web.app.page.WebPage;
+import com.github.freeacs.web.app.util.ACSLoader;
 import com.github.freeacs.web.app.util.Freemarker;
 import com.github.freeacs.web.app.util.StackTraceFormatter;
 import com.github.freeacs.web.app.util.WebProperties;
-import com.github.freeacs.web.app.util.ACSLoader;
 import freemarker.template.Configuration;
 import freemarker.template.Template;
 import freemarker.template.TemplateException;
