+++ web/src/main/java/com/github/freeacs/web/app/page/AbstractWebPage.java
@@ -4,9 +4,9 @@ import com.github.freeacs.common.util.NaturalComparator;
 import com.github.freeacs.dbi.*;
 import com.github.freeacs.web.Page;
 import com.github.freeacs.web.app.menu.MenuItem;
-import com.github.freeacs.web.app.security.AllowedUnittype;
 import com.github.freeacs.web.app.table.TableColor;
 import com.github.freeacs.web.app.util.*;
+import com.github.freeacs.web.security.AllowedUnittype;
 import freemarker.template.SimpleScalar;
 import freemarker.template.TemplateMethodModel;
 import freemarker.template.TemplateModel;
