+++ web/src/main/java/com/github/freeacs/web/app/page/certificates/CryptoPage.java
@@ -7,8 +7,8 @@ import com.github.freeacs.web.app.Output;
 import com.github.freeacs.web.app.input.InputDataRetriever;
 import com.github.freeacs.web.app.input.ParameterParser;
 import com.github.freeacs.web.app.page.AbstractWebPage;
-import com.github.freeacs.web.app.util.WebConstants;
 import com.github.freeacs.web.app.util.ACSLoader;
+import com.github.freeacs.web.app.util.WebConstants;
 import freemarker.template.TemplateMethodModel;
 import freemarker.template.TemplateModelException;
 import org.slf4j.Logger;
