+++ web/src/main/java/com/github/freeacs/web/app/page/heartbeat/HeartbeatsPage.java
@@ -6,9 +6,9 @@ import com.github.freeacs.web.app.Output;
 import com.github.freeacs.web.app.input.*;
 import com.github.freeacs.web.app.menu.MenuItem;
 import com.github.freeacs.web.app.page.AbstractWebPage;
+import com.github.freeacs.web.app.util.ACSLoader;
 import com.github.freeacs.web.app.util.SessionData;
 import com.github.freeacs.web.app.util.WebConstants;
-import com.github.freeacs.web.app.util.ACSLoader;
 
 import javax.sql.DataSource;
 import java.lang.reflect.InvocationTargetException;
