+++ web/src/main/java/com/github/freeacs/web/app/page/permissions/PermissionsPage.java
@@ -7,11 +7,11 @@ import com.github.freeacs.web.app.input.InputDataIntegrity;
 import com.github.freeacs.web.app.input.InputDataRetriever;
 import com.github.freeacs.web.app.input.ParameterParser;
 import com.github.freeacs.web.app.page.AbstractWebPage;
-import com.github.freeacs.web.app.security.WebUser;
+import com.github.freeacs.web.app.util.ACSLoader;
 import com.github.freeacs.web.app.util.SessionCache;
 import com.github.freeacs.web.app.util.SessionData;
 import com.github.freeacs.web.app.util.WebConstants;
-import com.github.freeacs.web.app.util.ACSLoader;
+import com.github.freeacs.web.security.WebUser;
 import freemarker.template.SimpleScalar;
 import freemarker.template.TemplateMethodModel;
 import freemarker.template.TemplateModelException;
