+++ web/src/main/java/com/github/freeacs/web/app/page/profile/ProfileDetailsPage.java
@@ -11,9 +11,9 @@ import com.github.freeacs.web.app.page.syslog.SyslogUtil;
 import com.github.freeacs.web.app.page.unittype.UnittypeParameterFlags;
 import com.github.freeacs.web.app.page.unittype.UnittypeParameterTypes;
 import com.github.freeacs.web.app.table.TableElementMaker;
+import com.github.freeacs.web.app.util.ACSLoader;
 import com.github.freeacs.web.app.util.SessionData;
 import com.github.freeacs.web.app.util.WebConstants;
-import com.github.freeacs.web.app.util.ACSLoader;
 
 import javax.sql.DataSource;
 import java.util.ArrayList;
