+++ web/src/main/java/com/github/freeacs/web/app/page/report/custom/DefaultRetriever.java
@@ -1,9 +1,9 @@
 package com.github.freeacs.web.app.page.report.custom;
 
+import com.github.freeacs.dbi.ACS;
 import com.github.freeacs.dbi.Group;
 import com.github.freeacs.dbi.Profile;
 import com.github.freeacs.dbi.Unittype;
-import com.github.freeacs.dbi.ACS;
 import com.github.freeacs.dbi.report.PeriodType;
 import com.github.freeacs.dbi.report.Report;
 import com.github.freeacs.dbi.report.ReportGenerator;
