+++ web/src/main/java/com/github/freeacs/web/app/page/search/SearchPage.java
@@ -7,8 +7,8 @@ import com.github.freeacs.web.Page;
 import com.github.freeacs.web.app.Output;
 import com.github.freeacs.web.app.input.*;
 import com.github.freeacs.web.app.page.AbstractWebPage;
-import com.github.freeacs.web.app.util.WebConstants;
 import com.github.freeacs.web.app.util.ACSLoader;
+import com.github.freeacs.web.app.util.WebConstants;
 import org.json.JSONArray;
 import org.json.JSONObject;
