+++ web/src/main/java/com/github/freeacs/web/app/page/trigger/TriggerHandler.java
@@ -6,8 +6,8 @@ import com.github.freeacs.web.app.input.Input;
 import com.github.freeacs.web.app.input.InputSelectionFactory;
 import com.github.freeacs.web.app.table.TableElement;
 import com.github.freeacs.web.app.table.TableElementMaker;
-import com.github.freeacs.web.app.util.SessionCache;
 import com.github.freeacs.web.app.util.ACSLoader;
+import com.github.freeacs.web.app.util.SessionCache;
 
 import javax.sql.DataSource;
 import java.sql.SQLException;
