+++ web/src/main/java/com/github/freeacs/web/app/page/unit/InspectionPage.java
@@ -5,8 +5,8 @@ import com.github.freeacs.dbi.util.SystemParameters;
 import com.github.freeacs.web.app.Output;
 import com.github.freeacs.web.app.input.*;
 import com.github.freeacs.web.app.page.AbstractWebPage;
-import com.github.freeacs.web.app.util.SessionCache;
 import com.github.freeacs.web.app.util.ACSLoader;
+import com.github.freeacs.web.app.util.SessionCache;
 
 import javax.sql.DataSource;
 import java.util.Arrays;
