--- web/src/main/java/com/github/freeacs/web/app/security/handlers/package.html
@@ -1,11 +1,0 @@
-<!DOCTYPE HTML PUBLIC "-//W3C//DTD HTML 3.2 Final//EN">
-<html>
-<head></head>
-<body bgcolor="white">
-
-<p>Contains authentication handlers for use in the user authentication system.</p>
-
-<p>Every authentication handler must implement AuthenticationInterface.</p>
-
-</body>
-</html>
