+++ web/src/main/java/com/github/freeacs/web/app/util/SessionData.java
@@ -2,8 +2,8 @@ package com.github.freeacs.web.app.util;
 
 import com.github.freeacs.dbi.Unit;
 import com.github.freeacs.dbi.UnitJob;
-import com.github.freeacs.web.app.security.AllowedUnittype;
-import com.github.freeacs.web.app.security.WebUser;
+import com.github.freeacs.web.security.AllowedUnittype;
+import com.github.freeacs.web.security.WebUser;
 
 import java.util.Collection;
 import java.util.Date;
