+++ web/src/main/java/com/github/freeacs/web/config/MvcConfig.java
@@ -1,5 +1,6 @@
 package com.github.freeacs.web.config;
 
+import com.github.freeacs.web.app.Main;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.web.servlet.config.annotation.ViewControllerRegistry;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
@@ -8,6 +9,7 @@ import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
 public class MvcConfig implements WebMvcConfigurer {
 
     public void addViewControllers(ViewControllerRegistry registry) {
+        registry.addRedirectViewController("/", Main.servletMapping);
         registry.addViewController("/login").setViewName("login");
     }
